# docs/swagger.yaml
basePath: /api
definitions:
...
  favorite.BatchCheckRequest:
    properties:
      recipe_ids:
        items:
          type: string
        type: array
    required:
    - recipe_ids
    type: object
  favorite.BatchCheckResponse:
    properties:
      favorited_ids:
        items:
          type: string
        type: array
    type: object
  favorite.CheckFavoriteResponse:
    properties:
      is_favorite:
        type: boolean
    type: object
  favorite.FavoriteListItem:
    properties:
      created_at:
        type: string
      recipe:
        $ref: '#/definitions/types.RecipeListItem'
      recipe_id:
        type: string
    type: object
  favorite.FavoriteListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/favorite.FavoriteListItem'
        type: array
      total:
        type: integer
    type: object
  favorite.FavoriteRequest:
    properties:
      recipe_id:
        type: string
    required:
    - recipe_id
    type: object
  favorite.FavoriteResponse:
    properties:
      created_at:
        type: string
      recipe_id:
        type: string
    type: object
  history.HistoryListItem:
    properties:
      recipe:
        $ref: '#/definitions/types.RecipeListItem'
      recipe_id:
        type: string
      viewed_at:
        type: string
    type: object
  history.HistoryListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/history.HistoryListItem'
        type: array
      total:
        type: integer
    type: object
  history.HistoryRequest:
    properties:
      recipe_id:
        type: string
    required:
    - recipe_id
    type: object
  history.HistoryResponse:
    properties:
      recipe_id:
        type: string
      viewed_at:
        type: string
    type: object
...
  types.RecipeListItem:
    properties:
      category:
        type: string
      description:
        type: string
      difficulty:
        type: integer
      id:
        type: string
      image_path:
        type: string
      name:
        type: string
      tags:
        $ref: '#/definitions/types.TagsGrouped'
      total_time_minutes:
        type: integer
    type: object
  types.TagsGrouped:
    properties:
      cuisines:
        items:
          type: string
        type: array
      flavors:
        items:
          type: string
        type: array
      scenes:
        items:
          type: string
        type: array
    type: object
...
host: localhost:18000
info:
  contact: {}
  description: 菜谱管理系统后端 API
  title: Choosy API
  version: "1.0"
paths:
...
  /api/home/hot:
    get:
      parameters:
      - default: 6
        description: 数量限制
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.RecipeListItem'
            type: array
      summary: 获取热门菜谱（按收藏数排序）
      tags:
      - home
  /api/home/recommend:
    get:
      parameters:
      - default: 4
        description: 数量限制
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.RecipeListItem'
            type: array
      summary: 获取推荐菜谱（随机）
      tags:
      - home
...
  /api/user/favorites:
    get:
      parameters:
      - description: 分类筛选
        in: query
        name: category
        type: string
      - description: 搜索关键词
        in: query
        name: search
        type: string
      - default: 20
        description: 限制数量
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/favorite.FavoriteListResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取收藏列表
      tags:
      - favorites
    post:
      consumes:
      - application/json
      parameters:
      - description: 收藏信息
        in: body
        name: favorite
        required: true
        schema:
          $ref: '#/definitions/favorite.FavoriteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/favorite.FavoriteResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 添加收藏
      tags:
      - favorites
  /api/user/favorites/{recipe_id}:
    delete:
      parameters:
      - description: 菜谱ID
        in: path
        name: recipe_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 取消收藏
      tags:
      - favorites
  /api/user/favorites/{recipe_id}/check:
    get:
      parameters:
      - description: 菜谱ID
        in: path
        name: recipe_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/favorite.CheckFavoriteResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 检查是否已收藏
      tags:
      - favorites
  /api/user/favorites/batch-check:
    post:
      consumes:
      - application/json
      parameters:
      - description: 菜谱ID列表
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/favorite.BatchCheckRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/favorite.BatchCheckResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 批量检查收藏状态
      tags:
      - favorites
  /api/user/history:
    delete:
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 清空浏览历史
      tags:
      - history
    get:
      parameters:
      - description: 分类筛选
        in: query
        name: category
        type: string
      - description: 搜索关键词
        in: query
        name: search
        type: string
      - default: 20
        description: 限制数量
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history.HistoryListResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取浏览历史列表
      tags:
      - history
    post:
      consumes:
      - application/json
      parameters:
      - description: 浏览记录信息
        in: body
        name: history
        required: true
        schema:
          $ref: '#/definitions/history.HistoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/history.HistoryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 添加浏览记录
      tags:
      - history
  /api/user/history/{recipe_id}:
    delete:
      parameters:
      - description: 菜谱ID
        in: path
        name: recipe_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除浏览记录
      tags:
      - history
...
securityDefinitions:
  Bearer:
    description: 输入 "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"